+++ emacs/giddyanne.el
@@ -3,7 +3,7 @@
 ;; Copyright (C) 2025 Joey
 
 ;; Author: Joey
-;; Version: 1.6.0
+;; Version: 1.6.1
 ;; Package-Requires: ((emacs "28.1"))
 ;; Keywords: tools, search, semantic
 ;; URL: https://github.com/mazziv/giddyanne